Move ring buffer to its own module
